Update references to reflect bogus mismatch counts in Stats.json
There's a bug in bcl2fastq 2.18 which caused the index mismatch counts
in Stats.json to be incorrect. The values in the xml files were correct. Since we
change to use the json file, the incorrect values are now used in the
refrences. After updating bcl2fastq, the values are actually correct
again.